backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssselect2==0.7.0
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
filelock==3.10.5
html5lib==1.1
idna==3.4
isort==5.12.0
lxml==4.9.2
mysqlclient==2.1.1
oscrypto==1.3.0
Pillow==9.4.0
platformdirs==3.2.0
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyMySQL==1.0.3
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
